% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breedgenr.R
\docType{package}
\name{breedgenr}
\alias{breedgenr}
\alias{breedgenr-package}
\title{breedgenr: A package for simulating breeding populations with non-parametric methods}
\description{
breedgenr uses real genotypic and phenotypic data from breeding programs
to generate the reference population
}
\details{
breedgenr implements three different non-parametric method:˚
randonForest ('RF', \code{rf}), support vector regression ('SVR' \code{svr})
and Nadaraya–Watson kernel regression ('NW', \code{nw})
}
