% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_population.R
\name{create_population}
\alias{create_population}
\title{Create or simulate a population based genotypic and phenotypic data}
\usage{
create_population(haplo = NULL, pheno = NULL, map = NULL,
  RR = NULL, h2 = NULL, simul = TRUE, method = c("RF", "SVR",
  "NW"), nb_gen = 15, nb_ind_gen = 2000, nb_ind_P0 = 500,
  folds = 10, n_rep = 10, nb_cores = 1, verbose = T, ...)
}
\arguments{
\item{haplo}{an array of dimension p, n and 2 or a matrix of dimension p, 2n (NAs not  allowed for NW method).}

\item{pheno}{a data frame or a matrix of dimension n and y.
The row names should identical to the column names of \code{haplo}.}

\item{map}{a data frame or a matrix of dimension p and 2.
The row names should identical to the row names of \code{haplo}.}

\item{RR}{a numeric value corresponding de the recombination rate.}

\item{h2}{a numeric value [0,1] or vector of length y.}

\item{simul}{logical scalar: TRUE (default) or FALSE.}

\item{method}{a character string indicating which method to use
for the calibration of the genotype-phenotype relationship:
"RF" (default), "SVR", or "NW".}

\item{nb_gen}{an integer indicating the number of generation
to simulate from the current population. Default = 15}

\item{nb_ind_gen}{an integer or a vector of the same length \code{nb_gen}
indicating the number of individual per generation. Default = 2000}

\item{nb_ind_P0}{an integer indicating the final number of individuals in the population}

\item{folds}{an integer for the number of folds. Default  = 10.}

\item{n_rep}{an integer for the number of replicates. Default  = 10.}

\item{nb_cores}{the number of cores to use for parallel execution. Default = 1.}

\item{verbose}{logical scalar: TRUE (default) or FALSE}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
A list :
 \itemize{
 \item{'haplo'}{ an array with the haplotype of the new population}
 \item{'geno'}{ a matrix with the genotype of the new population}
}
}
\description{
\code{create_population} ceate a population object based on genotypic and phenotypic data
}
\details{
#' This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{

pop <- create_population(
haplo = rice1_x,
pheno = rice1_y,
map = rice1_map,
RR = 244000,
h2 =rep(0.3, 3),
nb_gen = 5,
nb_ind_gen = 200,
nb_ind_P0 = 100,
mtry = 1000
)

}
\seealso{
\code{\link{simul_geno}}, \code{\link{simul_P0}}, \code{\link{rf}}, \code{\link{svr}}
}
