% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_fct.R
\name{nw}
\alias{nw}
\title{Title}
\usage{
nw(x, y, x_pred, h, kernel_type = c("rho", "gaussian", "aggregate"),
  opt_subset = NULL)
}
\arguments{
\item{x}{A data frame or a matrix of predictors (genotypes) of dimension n, p.}

\item{y}{A response vector (phenotype) of length n.}

\item{x_pred}{A data frame or a matrix (m, p) of the same predictors os \code{x} for a new set of observations.}

\item{h}{bandwith}

\item{kernel_type}{a character string indicating which method to use
for the kernel estimation: 'aggregate' (default), 'rho', 'gaussian'.}

\item{opt_subset}{not used}
}
\value{
A vector of predicted values of length m
}
\description{
Title
}
\examples{



}
