% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_fct.R
\name{nw_opt}
\alias{nw_opt}
\title{Optimisation of Nadaraya Watson regression parameters}
\usage{
nw_opt(x, y, kernel_type = c("rho", "gaussian", "aggregate"), h_grid,
  folds = 10, n_rep = 10, nb_cores = 1)
}
\arguments{
\item{x}{A data frame or a matrix of predictors (genotypes) of dimension n, p.}

\item{y}{A response vector (phenotype) of length n.}

\item{kernel_type}{a character string indicating which method to use
for the kernel estimation: 'aggregate' (default), 'rho', 'gaussian'.}

\item{folds}{an integer for the number of folds. Default = 10.}

\item{n_rep}{an integer for the number of replicates. Default = 10.}

\item{nb_cores}{the number of cores to use for parallel execution. Default = 1.}
}
\value{
A list
}
\description{
The \code{nw_opt} function performs a grid search to optimise
  the bandwith parameter for kernel estimation
}
\examples{
# May take some time
haplo <- rice1_x
x <- aperm(haplo, perm = c(2, 1, 3))
x <- x[, , 1] + x[, , 2]
y <- rice1_y[, 1]
system.time(
 h_opt <- nw_opt(
   x = x,
   y = y,
   kernel_type = 'aggregate',
   h_grid = kernlab::sigest(y ~ x, frac = 1),
   n_rep = 2,
   folds = 5,
   nb_cores = 2
 )
 )

}
