% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fct.R
\name{rf}
\alias{rf}
\title{Prediction using randomForest}
\usage{
rf(x, y, x_pred, ntree = 1000, mtry)
}
\arguments{
\item{x}{A data frame or a matrix of predictors (genotypes) of dimension n, p.}

\item{y}{A response vector (phenotype) of length n.}

\item{x_pred}{A data frame or a matrix (m, p) of the same predictors os \code{x} for a new set of observations.}

\item{ntree}{Number of trees (integer). Default is 1,000}

\item{mtry}{an integer indicating the number of variables randomly sampled at each split.
Default is the (rounded down) square root of the number variables.}
}
\value{
A vector of predicted values of length m
}
\description{
The \code{rf} function performs prediction using new genotypic data (x_pred)
}
\examples{


}
\seealso{
#'  \code{\link[ranger]{ranger}}, \code{\link[randomForest]{randomForest}}
}
