% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fct.R
\name{rf_opt}
\alias{rf_opt}
\title{Optimisation of randomForest parameters}
\usage{
rf_opt(x, y, ntree = 1000, mtry, folds = 10, n_rep = 10,
  nb_cores = 1)
}
\arguments{
\item{x}{A data frame or a matrix of predictors (genotypes) of dimension n, p.}

\item{y}{A response vector (phenotype) of length n.}

\item{ntree}{Number of trees (integer). Default is 1,000}

\item{mtry}{an integer indicating the number of variables randomly sampled at each split.
Default is the (rounded down) square root of the number variables.}

\item{folds}{an integer for the number of folds. Default = 10.}

\item{n_rep}{an integer for the number of replicates. Default = 10.}

\item{nb_cores}{the number of cores to use for parallel execution. Default = 1.}
}
\value{
The model
}
\description{
The \code{rf_opt} function performs a grid search to optimise
  the mtry parameter mtry for randomForest method.
}
\examples{



}
\seealso{
\code{\link[caret]{train}}, \code{\link[ranger]{ranger}}, \code{\link[randomForest]{randomForest}}
}
