% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_geno.R
\name{simul_geno}
\alias{simul_geno}
\title{Generate a reference population (P0) from real genotypic and phenotypic data.}
\usage{
simul_geno(haplo, map, RR, nb_gen = 15, nb_ind_gen = 2000,
  nb_ind_P0 = 500, nb_cores = 1, verbose = T)
}
\arguments{
\item{haplo}{an array of dimension p, n and 2 or a matrix of dimension p, 2n (NAs not  allowed for NW method).}

\item{map}{a data frame or a matrix of dimension p and 2.
The row names should identical to the row names of \code{haplo}.}

\item{RR}{a numeric value corresponding de the recombination rate.}

\item{nb_gen}{an integer indicating the number of generation
to simulate from the current population. Default = 15}

\item{nb_ind_gen}{an integer or a vector of the same length \code{nb_gen}
indicating the number of individual per generation. Default = 2000}

\item{nb_ind_P0}{an integer indicating the final number of individuals in the population}

\item{nb_cores}{the number of cores to use for parallel execution. Default = 1.}

\item{verbose}{logical scalar: TRUE (default) or FALSE}
}
\value{
A list with two object:
  * an array containing the haplotypes
  * a matrix with the genotypes
}
\description{
\code{simul_geno} compute...
}
\examples{

# Takes time!
system.time(
  P0 <- simul_geno(
    haplo = rice1_x,
    map = rice1_map,
    RR = 244000,
    nb_ind_gen = 1000,
    nb_gen = 2,
    nb_ind_P0 = 200,
  )
 )

dim(P0$sim_haplo)
dim(P0$sim_geno)


}
