% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_P0.R
\name{simul_P0}
\alias{simul_P0}
\title{Simulate a population}
\usage{
simul_P0(haplo, pheno, map, RR, h2, nb_gen = 15, nb_ind_gen = 2000,
  nb_ind_P0 = 500, folds = 10, n_rep = 10, method = c("RF", "SVR",
  "NW"), nb_cores = 1, verbose = T, ...)
}
\arguments{
\item{haplo}{an array of dimension p, n and 2 or a matrix of dimension p, 2n (NAs not  allowed for NW method).}

\item{pheno}{a data frame or a matrix of dimension n and y.
The row names should identical to the column names of \code{haplo}.}

\item{map}{a data frame or a matrix of dimension p and 2.
The row names should identical to the row names of \code{haplo}.}

\item{RR}{a numeric value corresponding de the recombination rate.}

\item{h2}{a numeric value [0,1] or vector of length y.}

\item{nb_gen}{an integer indicating the number of generation
to simulate from the current population. Default = 15}

\item{nb_ind_gen}{an integer or a vector of the same length \code{nb_gen}
indicating the number of individual per generation. Default = 2000}

\item{nb_ind_P0}{an integer indicating the final number of individuals in the population}

\item{folds}{an integer for the number of folds. Default  = 10.}

\item{n_rep}{an integer for the number of replicates. Default  = 10.}

\item{method}{a character string indicating which method to use
for the calibration of the genotype-phenotype relationship:
"RF" (default), "SVR", or "NW".}

\item{nb_cores}{the number of cores to use for parallel execution. Default = 1.}

\item{verbose}{logical scalar: TRUE (default) or FALSE}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
A list :
 *
 *
}
\description{
\code{simul_P0} simulate a population based on genotypic and phenotypic data
}
