% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fct.R
\name{svr}
\alias{svr}
\title{Prediction using support vector regression with gaussian kernel}
\usage{
svr(x, y, x_pred, C, sigma)
}
\arguments{
\item{x}{A data frame or a matrix of predictors (genotypes) of dimension n, p.}

\item{y}{A response vector (phenotype) of length n.}

\item{x_pred}{A data frame or a matrix (m, p) of the same predictors os \code{x} for a new set of observations.}

\item{C}{a numeric value for the soft margin parameter}

\item{sigma}{a numeric value indicating the inverse kernel width for the gaussian kernel}
}
\value{
A vector of predicted values of length m
}
\description{
The \code{svr} function performs prediction using new genotypic data (x_pred)
}
