% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fct.R
\name{svr_opt}
\alias{svr_opt}
\title{Optimisation of support vector regression with gaussian kernel}
\usage{
svr_opt(x, y, C, sigma = NULL, folds = 10, n_rep = 10,
  nb_cores = 1)
}
\arguments{
\item{x}{A data frame or a matrix of predictors (genotypes) of dimension n, p.}

\item{y}{A response vector (phenotype) of length n.}

\item{C}{a numeric value for the soft margin parameter}

\item{sigma}{a numeric value indicating the inverse kernel width for the gaussian kernel}

\item{folds}{an integer for the number of folds. Default = 10.}

\item{n_rep}{an integer for the number of replicates. Default= 10.}

\item{nb_cores}{the number of cores to use for parallel execution. Default = 1.}
}
\value{
The model
}
\description{
The \code{svr_opt} function performs a grid search to optimise
  the C and sigma parameters for support vector regression.
  The \code{\link[caret]{train}} function is used.
}
